pyproject.toml
[tool.poetry]
name = "spacedice"
version = "0.1.0"
description = ""
authors = ["E1000"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tk = "^0.1.0"
flask = "^3.0.3"
kivy = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
